projects/monitoring/1 - otel stdout exporter/pyproject.toml
[project]
name = "otel-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.118.3",
    "opentelemetry-distro>=0.58b0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.37.0",
    "opentelemetry-instrumentation-logging>=0.58b0",
    "uvicorn>=0.37.0",
]
